requirements.txt
pandas~=2.2.3
python-dotenv~=1.1.0
SQLAlchemy~=2.0.40
yfinance~=0.2.58
matplotlib~=3.10.1
seaborn~=0.13.2
ipywidgets~=8.1.7